change Vehicle.capacity type from double to int

+++ main/java/org/matsim/vsp/ev/dvrp/EvDvrpVehicle.java
@@ -56,7 +56,7 @@ public class EvDvrpVehicle implements Vehicle {
 	}
 
 	@Override
-	public double getCapacity() {
+	public int getCapacity() {
 		return vehicle.getCapacity();
 	}
